# Performance improvements and search optimisation for select dropdowns