--- a/.github/workflows/deploy.yml
+++ b/.github/workflows/deploy.yml
@@ -18,6 +18,7 @@ jobs:
     # The type of runner that the job will run on
     runs-on: ubuntu-latest
     needs: build
+    environment: production
 
     # Steps represent a sequence of tasks that will be executed as part of the job
     steps: