--- a/.gitignore
+++ b/.gitignore
@@ -13,3 +13,4 @@ storage/signups/*
 _ide_helper.php
 _ide_helper_models.php
 build/prod/db/*.sql
+build/*dev