--- a/app/Http/Controllers/ResultsController.php
+++ b/app/Http/Controllers/ResultsController.php
@@ -245,9 +245,9 @@ class ResultsController extends Controller
 
     public function uploadResults()
     {
-        $season = Season::where('status', '<', 2)->get();
+        $season = Season::where('status', '<', 2)->get()->load('series');
         $points = Points::all();
-        $tracks = Circuit::select('id', 'name')->get();
+        $tracks = Circuit::select('id', 'name', 'series')->get()->load('series');
         $constructor = Constructor::select('id', 'name')->get();
         $driver = Driver::select('id', 'name')->get();