--- a/resources/views/standings/upload.blade.php
+++ b/resources/views/standings/upload.blade.php
@@ -435,6 +435,9 @@
                 value: 'DSQ'
             }
         ];
+
+        console.log(season)
+        console.log(tracks)
         
         $('#editScreen').addClass('hidden');
         $('#serverResponseScreen').addClass('hidden');
@@ -985,7 +988,7 @@
             let dispValue;
             
             for(let i = 0; i < selectInpFields[key].data.length; i++) {
-                dispValue = selectInpFields[key].data[i].name;
+                dispValue = `${selectInpFields[key].data[i].name} (${selectInpFields[key].data[i].series.code.toUpperCase()})`;
                 
                 if(selectInpFields[key].data[i].id !== selectInpFields[key].upload) {
                     select += `<option value=${selectInpFields[key].data[i].id}>${dispValue}</option>`;
@@ -1189,14 +1192,14 @@
 
         for(let y = 0; y < selectInpFields.driverID.originalData.length; y++) {
             if(selectInpFields.driverID.originalData[y].id === selectInpFields.driverID.upload) {
-                driverDispVal = selectInpFields.driverID.originalData[y].name;
+                driverDispVal = `${selectInpFields.driverID.originalData[y].name} (${selectInpFields.driverID.originalData[y].id})`;
                 driverSelect += `<option selected value=${selectInpFields.driverID.originalData[y].id}>${driverDispVal}</option>`;
             }
         }
 
         // Populate driver dropdown without the selected driver from the available drivers
         for(let x = 0; x < selectInpFields.driverID.data.length; x++) {
-            driverDispVal = selectInpFields.driverID.data[x].name;         
+            driverDispVal = `${selectInpFields.driverID.data[x].name} (${selectInpFields.driverID.data[x].id})`;         
             driverSelect += `<option value=${selectInpFields.driverID.data[x].id}>${driverDispVal}</option>`;
         }
 
@@ -1208,7 +1211,7 @@
         let constructorDispVal;
         
         for(let x = 0; x < selectInpFields.constructorID.data.length; x++) {
-            constructorDispVal = `${selectInpFields.constructorID.data[x].id} - ${selectInpFields.constructorID.data[x].name}`;
+            constructorDispVal = `${selectInpFields.constructorID.data[x].name} (${selectInpFields.constructorID.data[x].id})`;
 
             if(selectInpFields.constructorID.data[x].id !== selectInpFields.constructorID.upload) {
                 constructorSelect += `<option value=${selectInpFields.constructorID.data[x].id}>${constructorDispVal}</option>`;
@@ -1526,7 +1529,7 @@
             let constructorCol = document.getElementById(`constructorSelect${x}`);
 
             for(let y = 0; y < supportingVariables.availableConstructors.length; y++) {
-                constructorDispVal = `${supportingVariables.availableConstructors[y].id} - ${supportingVariables.availableConstructors[y].name}`;
+                constructorDispVal = `${supportingVariables.availableConstructors[y].name} (${supportingVariables.availableConstructors[y].id})`;
 
                 if(supportingVariables.availableConstructors[y].id !== json.results[x].constructor_id) {
                     constructorSelect += `<option value=${supportingVariables.availableConstructors[y].id}>${constructorDispVal}</option>`;
@@ -2040,13 +2043,13 @@
 
             for(let y = 0; y < driver.length; y++) {
                 if(driver[y].id === json.results[x].driver_id) {
-                    driverDispVal = driver[y].name;
+                    driverDispVal = `${driver[y].name} (${driver[y].id})`;
                     driverSelect += `<option selected value=${driver[y].id}>${driverDispVal}</option>`;
                 }
             }
 
             for(let y = 0; y < supportingVariables.availableDrivers.length; y++) {
-                driverDispVal = supportingVariables.availableDrivers[y].name;  
+                driverDispVal = `${supportingVariables.availableDrivers[y].name} (${supportingVariables.availableDrivers[y].id})`;  
                 driverSelect += `<option value=${supportingVariables.availableDrivers[y].id}>${driverDispVal}</option>`;
             }
 
@@ -3285,7 +3288,7 @@
             checkRaceTimeMatchesWithStatus(updatedJSONFromTable, jsonResultsDetailsStore, supportingVariables, i);
             checkAndMonitorResultsData(json, points, driver, supportingVariables.availableConstructors, status, regexValidationStrings, jsonTrackDetailsStore, jsonResultsDetailsStore, additionalDetailsStore, supportingVariables, i);
             serialiseRowReorderControls(jsonResultsDetailsStore, additionalDetailsStore, supportingVariables, i);
-            
+
             openResultsMoreDetailsOverlay(json, additionalDetailsStore, supportingVariables, i);
             
             checkGridValueGreaterThanArraySize(jsonResultsDetailsStore, supportingVariables);