## What is the issue?
On selecting a different driver from the driver dropdown, the webpage takes a considerable amount of time to assign the selected driver in the field, freezing the webpage in the process.

A similiar instance of this issue was found while adding rows into the results table.
The solution found was exactly the same.

This causes unecessary delay in the workflow and a negative experience for the user.

## Troubleshooting Steps:
The approach taken to resolve this issue are as follows:
1. Identify the function being called on selection of a driver from the dropdown
1. Measure time taken for subsequent function calls within the wrapper function
1. Navigate to the function with significantly high call time and measure time for each operation
1. Reduce operation times and bring down the overall call time to acceptable range

### 1. Identifying function call on selection
On reviewing the call trace, it was found that `checkAndMonitorDriverId` was called when changing/selecting a driver from the dropdowns.

Within the above function, the 'change' event listener catches any changes made to the driver select field and carries out the job of reassigning current driver and repopulating all driver dropdowns.

Here are the function calls within the listener:

- Filter drivers.
- Repopulate all driver dropdowns.
- Check for duplicate driver selected.
- Check for invalid driver selection.
- Check for errors in all editable table cells.

### 2. Measuring call times of all functions within the wrapper
For measuring call times, `console.time(label)` was used before each function call to start the timer and `console.timeEnd(label)`, on the next line after the call.

Call times were measured after changing driver in position 1 to **The-Real-SD**, of race result with ID of **900**.

On measuring call times of all functions within the listener, the following results were seen:

    Listener total : 6467.94921875 ms
        - Filter : 0.035888671875 ms
        - Repopulation : 6429.35107421875 ms
        - Duplicate : 36.407958984375 ms
        - Warning : 0.1708984375 ms
        - Error : 1.417236328125 ms

From the above call times, it was clear that repopulating all driver dropdowns was taking too long, compared to other function calls.

### 3. Measuring time for each operation
Mentioned below are the main operations inside the targeted function:

- Running the check and repopulation of options for all dropdowns.
- Checking the presence of selected driver in the stored driver list and if present, assigning the driver as selected option.
- Repopulating options of the current dropdown with filtered available drivers list.

Measurement of operation times was done by inserting `console.time(label)` before the start of each code block, for each operation and `console.timeEnd(label)` after each block.

The following observations were made:

    Overall Call Time: 5842.717041015625 ms
        - Assigning current driver: 0.005126953125 ms
        - Repopulating all options: 900.987060546875 ms

### 4. Reducing operation time and subsequent call times
Looking through the code, it was noted that the assignment of all options into current driver select node was done incorrectly.

Instead of replacing the HTML content of the select node, at the end of the creation of the options list, the HTML content was changed on every addition to the options list. This meant that the options inside the driver dropdown, were changing as many times as there were drivers in the filtered available drivers list, which is highly time consuming.

Moving the content replacement after the options list creation bore the following call times:

    Overall Call Time: 6429.35107421875 ms
        - Assigning current driver: 0.017822265625 ms
        - Repopulating all options: 2.287841796875 ms

## Final Result:
A summarised call time improvement can be found below:

|  Wrapper Fn  |  Called Fn  |  Operation  |  Time Before(ms)  |  Time After(ms)  | 
|  :--------  |  :-------  |  :--------  |  -------:  |  --------:  |
|  Change Listener |    |    |  6467.94921875  |  58.427001953125  |
|    |  Filter  |    |  0.035888671875  |  0.065185546875  |
|    |  Repopulation  |    |  6429.35107421875  |  28.956787109375  |
|    |    |  Assigning current driver  |  0.005126953125  |  0.005126953125  |
|    |    |  Repopulating all options  |  900.987060546875  |  2.684814453125  |
|    |  Duplicate  |    |  36.407958984375  |  27.467041015625  |
|    |  Warning  |    |  0.1708984375  |  0.173095703125  |
|    |  Error  |    |  1.417236328125  |  0.89501953125  |

Through this change, the repopulation of all driver dropdowns take **99.55%** less time.